UPDATE: 게시글 수정 부분 수정

+++ src/main/java/com/example/blog/domain/Post.java
@@ -20,7 +20,7 @@ public class Post extends BaseTime {
     @Id
     @Column(name="post_id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long postId;
+    private Integer postId;
 
     public Post(User user, String author,String title, String content) {
         this.user = user;
